feat: Add mapToSource and mapFromSource helpers
Add helper functions to map a row number to and from a source model.
This avoids the need of creating a QModelIndex first when mapping
indexes between the proxy model and the source model.
Mapping with a QModelIndex is still possible, the return type of the
helpers will be the same as the one of the input parameter, a QModelIndex
or a row number.

Fix #23