CMake: Add build options BUILD_SHARED_LIB and BUILD_OBJECT_LIB
SortFilterProxyModel is an awesome piece of code and we use it in several
projects. This is the reason for this patch introducing shared library support.
By doing so we:

* Save memory by using shared library - at least RAM when running multiple
  applications using SortFilterProxyModel
* Save build time: SortFilterProxyModel is build once for all

To remain compatible with previous behaviour user can select what to do by cmake
options:

BUILD_OBJECT_LIB:
(Default ON): Same as before: Add sources to your project and build obj-files
that can be linked to your binary. It was taken care that projects behaves same
as before (tested) - no need for develepers to make changes in their project.

BUILD_SHARED_LIB: Build and install a shared library so that other projects can
use it. This is the way packagers prefer deploying code.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>