Add SortFilterProxyModel::registerQml() for environments linking with --as-needed
There are build environments that link their binaries with --as-needed for good
reasons [1].

This causes trouble here too: The application linking against
SortFilterProxyModel has no reference to a symbol within SortFilterProxyModel
so the library will simply be skipped and not loaded at runtime causing:

| qrc:/qml/FaFilter.qml:2:1: module "SortFilterProxyModel" is not installed

Linker with --as-needed behave similar as linking against static libraries and
Qt documatation [2] warns explicitly not to use Q_COREAPP_STARTUP_FUNCTION when
linking against static libraries.

To get around a stateless class SortFilterProxyModel (luckily this name was not
used yet) was created, that just contains the static method registerQml(). To
avoid multile calls on environemts that do not link with --as-needed all
register functions ensure to be called once only by a variable wasRegistered.

In the following patch Readme.md will updated and will recommend highly to call
SortFilterProxyModel::registerQml() - who knows which environment your
application will be build in?

[1] https://wiki.gentoo.org/wiki/Project:Quality_Assurance/As-needed
[2] https://doc.qt.io/qt-5/qcoreapplication.html#Q_COREAPP_STARTUP_FUNCTION

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>